requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
requests
python-dotenv